Refactor EditTestimonialPage to utilize EditTestimonialForm component for improved code organization and async data handling.
